Adopt typesafe logging for grid
